fix team ID/draft order discrepancy.
Last year the team IDs and draft orders were the same, so we were using the same integer for everything.  This year that's no longer the case.

Client-side everything is still expressed in terms of draft order.  Server-side we no longer deal with integers directly except when reading/writing to the database and serializing/deserializing to the wire; we introduce two separate wrapper classes for draft order and team ID so that we're explicit about which we're dealing with at all times.